Initial commit based on Jetson Linux 36.4.3
Added custom version of nvidia-kernel-oot recipe to include adjusted
CSI2 drivers for Alvium.
Added bbappend file for avt_csi2 driver to correctly setup the build for
tegra systems.
New machine "jetson-orin-nano-devkit-alvium-nvme" based on the standard
machine "jetson-orin-nann-devkit-nvme" which supports two Alvium cameras
out of the box. When using this machine configuration the patched
version of the nvidia-kernel-oot recipe is automatically selected and
the required camera device tree overlay is applied.